requirements/base.txt
# Basic stuff
django==1.9.7

# Templates
django-appconf==1.0.1
django-compressor==1.5

# Database
#psycopg2==2.6.1
#MySQL-python==1.2.5

# Admin styling
django-flat-theme==1.1.1

# Bcrypt passwords
# bcrypt==1.1.1

django-crispy-forms
django-extensions
werkzeug
untangle
xmltodict
django-datatable
crispy-forms-foundation
python-dateutil
django-select2
